requirements.txt
# ─────────────────────────────────────────────
# EduGen AI Agent — Requirements
# ─────────────────────────────────────────────

# ── Core LLM Framework ───────────────────────
langchain==0.2.16
langchain-community==0.2.16
langchain-openai==0.1.23
langchain-google-genai==1.0.10

openai==1.45.0
tiktoken==0.7.0

# ── UI ──────────────────────────────────────
streamlit==1.38.0
streamlit-extras==0.4.3
streamlit-option-menu==0.3.12

# ── PDF Processing ──────────────────────────
PyPDF2==3.0.1
pdfplumber==0.11.4

# ── Google APIs ─────────────────────────────
google-cloud-texttospeech==2.17.2
google-api-python-client==2.143.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

# ── YouTube ─────────────────────────────────
youtube-transcript-api==0.6.2
pytube==15.0.0

# ── Document Processing ─────────────────────
python-docx==1.1.2
notion-client==2.2.1

# ── Audio ───────────────────────────────────
gTTS==2.5.3

# ── HTTP / Scraping ─────────────────────────
requests==2.32.3
httpx==0.27.2
aiohttp==3.10.5
beautifulsoup4==4.12.3
lxml==5.3.0

# ── Utilities ───────────────────────────────
python-dotenv==1.0.1
pydantic==2.8.2
tqdm==4.66.5
loguru==0.7.2
tenacity==8.2.2
python-multipart==0.0.9

# ── Testing ─────────────────────────────────
pytest==8.3.2
pytest-asyncio==0.23.8